+++ Changes.md
@@ -15,6 +15,7 @@ Fixes
 -----
 
 - Cycles : Fixed rendering of shaders written with `IECOREUSD_WRITE_CONFORMANT_OSL_SHADERS=1`.
+- NodeEditor : Fixed flicker when changing the node being viewed.
 
 1.7.1.0 (relative to 1.7.0.0)
 =======